Use the new RedirectThread + include system PYTHONPATH
